% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_mpra_or_stop}
\alias{.is_mpra_or_stop}
\title{Validate MPRA Object Class}
\usage{
.is_mpra_or_stop(object)
}
\arguments{
\item{object}{An R object to be checked. It is expected to be an instance of the `MPRASet` class.}
}
\value{
This function does not return a value. If the class of `object` is not `MPRASet`, an error is thrown.
}
\description{
This function checks whether the provided object is of class `MPRASet`. If the object does not belong to this class, an error is raised.
}
\details{
The function performs a class check on the input `object`. If the object is not of class `MPRASet`, it stops the execution and raises an informative error message, specifying the actual class of the object.
This ensures that subsequent operations requiring an `MPRASet` object are only executed on valid inputs.
}
\examples{
\dontrun{
# Assuming `mpra_object` is an MPRA object
.is_mpra_or_stop(mpra_object) # This will pass if mpra_object is of class MPRASet
}

}
\keyword{internal}
