% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.onLoad}
\alias{.onLoad}
\title{Initialize Package Environment on Load}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{A character string representing the library name. This is passed by the R environment during package loading.}

\item{pkgname}{A character string representing the package name. This is passed by the R environment during package loading.}
}
\value{
This function does not return a value. It is executed for side effects during the loading of the package.
}
\description{
This function is called automatically when the package is loaded. It overrides the `compute_logratio` function in the `mpra` namespace with a custom implementation.
}
\details{
The function is invoked when the package is loaded using `library()`. It assigns a custom implementation of the `compute_logratio` function to the `mpra` namespace, ensuring that the modified version of `compute_logratio` is used in the context of the loaded package.

This is useful when you want to override functions in a package namespace at the time of loading, without modifying the original package source code.
}
\examples{
\dontrun{
# This function is executed automatically when the package is loaded
# No direct call is needed.
}

}
\keyword{internal}
