% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{.pivot_df}
\alias{.pivot_df}
\title{Pivot a data frame}
\usage{
.pivot_df(df, id_column_name = "variant_id", allele_column_name = NULL, type)
}
\arguments{
\item{df}{A data frame to be pivoted}

\item{id_column_name}{Character string specifying the name of the ID column. Default is "variant_id"}

\item{allele_column_name}{Character string specifying the name of the allele column. Default is NULL}

\item{type}{Character string specifying the type of values to pivot}
}
\value{
A pivoted data frame with rows named by the ID column and columns prefixed with "bc"
}
\description{
This function pivots a data frame, creating a wider format based on specified columns.
}
\examples{
df <- data.frame(
    variant_id = c("rs1", "rs1", "rs2", "rs2"),
    allele = c("ref", "alt", "ref", "alt"),
    count = 1:4
)
pivoted_df <- .pivot_df(df,
    id_column_name = "variant_id",
    allele_column_name = "allele", type = "count"
)

}
\keyword{internal}
