% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpra_set.R
\name{MPRASet}
\alias{MPRASet}
\title{Create an MPRASet Object}
\usage{
MPRASet(label = new("character"), ...)
}
\arguments{
\item{label}{A character vector containing the labels for each sequence. The length of the `label` vector must match the number of elements in the object, and it will be assigned to the `label` column in the row data of the MPRASet. Defaults to an empty character vector, meaning no labels will be applied unless specified.}

\item{...}{Additional arguments passed to the `mpra::MPRASet` constructor, allowing for further customization of the object (e.g., data matrices, metadata).}
}
\value{
An `MPRASet` object with or without labels, depending on the input.
}
\description{
This function initializes a new MPRASet object from the `mpra` package, optionally adding sequence labels to the object. The `label` parameter allows you to assign specific labels to each sequence in the dataset.
}
\details{
The function creates a new `MPRASet` object, which is a container for MPRA data. If the `label` vector is provided, it will be assigned to the `label` field of the row data. The `label` will be matched to the element identifiers (EIDs) of the object.

If no `label` is provided, the object will be returned without labels, and the user can assign them later using the `rowData` function.
}
\examples{
\dontrun{
# Create an MPRASet with sequence labels (other_args need to be defined before running this example)
mpra_data <- MPRASet(label = c("seq1", "seq2", "seq3"), other_args)
}

}
