% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{create_rna_df}
\alias{create_rna_df}
\title{Generate RNA count specific dataframe}
\usage{
create_rna_df(df, id_column_name = "variant_id", allele_column_name = NULL)
}
\arguments{
\item{df}{Data frame with 'name' column and the barcode and count information for each sequence}

\item{id_column_name}{Character string specifying the name of the ID column. Default is "variant_id"}

\item{allele_column_name}{Character string specifying the name of the allele column. Default is NULL}
}
\value{
Data frame with variant_id, allele, Barcode, and count columns
}
\description{
Generate RNA count specific dataframe
}
