% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{create_var_df}
\alias{create_var_df}
\title{Match count information and variant alleles to create a variant data frame}
\usage{
create_var_df(df, map_df)
}
\arguments{
\item{df}{Data frame with 'name' column and the barcode and count information for each sequence}

\item{map_df}{Data frame with 'ID', 'REF', and 'ALT' columns}
}
\value{
Data frame with variant_id, allele, Barcode, and count columns
}
\description{
Match count information and variant alleles to create a variant data frame
}
\examples{
df <- data.frame(name = c("ref1", "ref2", "alt1", "alt2"), Barcode = 1:4, dna_count= 2:5, rna_count = 10:13)
map_df <- data.frame(ID = c("rs1", "rs2"), REF = c("ref1", "ref2"), ALT = c("alt1", "alt2"))
result <- create_var_df(df, map_df)
head(result)
}
