% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{downsample_barcodes}
\alias{downsample_barcodes}
\title{Downsample barcodes in a data frame}
\usage{
downsample_barcodes(df, id_column_name = "name", percentile = 0.95)
}
\arguments{
\item{df}{A data frame containing barcode data.}

\item{id_column_name}{Character string specifying the name of the column
used for grouping. Default is "name".}

\item{percentile}{Numeric value between 0 and 1 specifying the percentile
to use for downsampling. Default is 0.95.}
}
\value{
A data frame with downsampled barcodes. If the specified id_column_name
  doesn't exist, returns the original data frame with a warning.
}
\description{
This function downsamples the number of barcodes for each group in a data frame
to a specified percentile of the distribution of barcode counts across groups.
}
\details{
The function first checks if an "allele" column exists in the data frame.
If it does, downsampling is performed within each combination of id and allele.
If not, downsampling is performed within each id group.

The number of barcodes kept for each group is determined by the specified
percentile of the distribution of barcode counts across all groups.
}
\examples{
df <- data.frame(
    name = rep(c(rep("A_seq", 2), rep("B_seq", 2), rep("C_seq", 6)), each = 2),
    barcode = paste0("barcode_", 1:20),
    allele = rep(c("ref", "alt"), 10),
    count = runif(20)
)

downsampled_df <- downsample_barcodes(df, id_column_name = "name", percentile = 0.8)
}
