% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpra_set.R
\name{getLabel}
\alias{getLabel}
\title{Extract Sequence Labels from an MPRA Object}
\usage{
getLabel(object)
}
\arguments{
\item{object}{An MPRA object, typically created using the `MPRASet` function, containing sequence data and labels. The object must contain a `label` column in its row data.}
}
\value{
A character vector of sequence labels from the row data of the MPRA object.
}
\description{
This function retrieves the sequence labels stored in the `label` field of the row data from an MPRA object. It is a simple accessor function designed to extract the label information for sequences in the dataset.
}
\details{
The function checks that the input object is a valid MPRA object using `.is_mpra_or_stop`. It then extracts the sequence labels from the `label` field of the row data. This field is expected to be a vector of labels for each sequence in the object, which is typically assigned during the creation of the object.

If the object does not contain a `label` field, this function will raise an error.
}
\examples{
\dontrun{
# Assuming `mpra_object` is an MPRA object with sequence labels
labels <- getLabel(mpra_object)
}

}
