% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{mpra_treat}
\alias{mpra_treat}
\title{Differential Logratio Testing for MPRA Data}
\usage{
mpra_treat(
  mpra_fit,
  percentile = NULL,
  neg_label,
  test_label = NULL,
  side = "both"
)
}
\arguments{
\item{mpra_fit}{A data frame containing MPRA fit data with at least two columns: \code{logratio} (numeric values representing log ratios for each entry) and \code{label} (categorical labels for groups).}

\item{percentile}{Numeric, the threshold percentile to use for filtering significant shifts in logratio values. If \code{NULL}, defaults to 0.95.}

\item{neg_label}{A character string specifying the label to use as the "negative control" group. This parameter is required.}

\item{test_label}{A character string specifying the label for the test group to compare against the negative control. If \code{NULL}, all labels except the negative label are considered.}

\item{side}{Character string specifying the direction of the test. Accepts \code{"both"} (default), \code{"right"}, or \code{"left"} for two-sided, right-tailed, or left-tailed tests, respectively.}
}
\value{
A data frame containing rows that meet the threshold criterion, with the following columns:
\itemize{
  \item \code{logFC}: Adjusted log fold change.
  \item \code{AveExpr}: Adjusted average expression.
  \item \code{name}: Sequence or gene names as extracted from row names.
}
}
\description{
Applies a thresholding approach to identify significantly shifted logratios in test groups relative to a negative control.
This function allows for side-specific testing, with options to test logratios that are significantly higher, lower, or both, compared to the negative group.
}
\details{
This function shifts the \code{logratio} values in \code{mpra_fit} by subtracting the mean of the negative control group (specified by \code{neg_label}) to standardize comparisons.
It then applies a threshold using the specified \code{percentile} on the logratios for the negative group, treating this as the cut-off for identifying shifts in the test group.
Side-specific testing is applied based on the \code{side} parameter, with options to filter for significant logratios in the positive (right), negative (left), or both directions.
This means if you run a two-sided test the number of results is only missing the rows of the "neg_label" sequences.
}
\note{
Requires the \code{edgeR} package for the \code{treat} function used in threshold testing.
}
\examples{
\dontrun{
mpra_data <- data.frame(logratio = rnorm(100), label = sample(c("control", "test"), 100, replace = TRUE))
significant_results <- mpra_treat(mpra_data, percentile = 0.95, neg_label = "control", test_label = "test", side = "both")
}

}
